requirements.txt
appdirs==1.4.3
asgiref==1.0.0
attrs==16.3.0
autobahn==0.17.2
Automat==0.5.0
channels==1.0.3
constantly==15.1.0
daphne==1.0.3
Django==2.0
incremental==16.10.1
oauthlib==2.0.1
packaging==16.8
psycopg2==2.8.6
psycopg2-binary==2.9.1
pyparsing==2.2.0
python-decouple==3.4
pytz==2021.1
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
Twisted==17.1.0
txaio==2.6.1
typing==3.5.3.0
zope.interface==4.3.3